RobotFlask/requirements.txt
azure-iothub-service-client==1.1.0.15
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pip==9.0.1
pymongo==3.5.1
requests==2.18.4
setuptools==28.8.0
urllib3==1.22
Werkzeug==0.12.2
